pyproject.toml
[project]
name = "db-tools-mcp"
version = "0.2.0"
description = "MCP server for querying SQL Server and Snowflake schema metadata"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastmcp>=3.0.2",
    "loguru>=0.7.3",
    "pyodbc>=5.3.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.47",
]

[project.scripts]
db-tools-mcp = "db_tools.server:main"
db-tools-refresh = "db_tools._extractor:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/db_tools"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
